refactor(cli): split native and builder entrypoints
Move the native trtmc executable source out of examples and split its argument parser into a testable src/cli module. Rename the Python builder CLI implementation to build_cli.py while keeping tensorrt_model_connect.cli as a compatibility alias for older imports.

Also wire trtmc detect through the native CLI, share native binary discovery with the Python pipeline wrapper, and update docs/tests to describe the new layout.
