+++ tensorrt_model_connect/tensorrt_model_connect/__main__.py
@@ -1,5 +1,5 @@
 """Allow `python -m tensorrt_model_connect` for the `trtmc build` bridge."""
 
-from .cli import main
+from .build_cli import main
 
 main()
