+++ tests/builder/test_cli_coverage.py
@@ -1,4 +1,4 @@
-"""Coverage-focused tests for CLI control flow in tensorrt_model_connect.cli.
+"""Coverage-focused tests for CLI control flow in tensorrt_model_connect.build_cli.
 
 Trace: ARCH-ENG-001, UD-ENG-02
 Intent: Validate CLI control flow branches including version lookup fallbacks, inspect output formatting, build dispatch, and error handling.
@@ -21,7 +21,17 @@ import pytest
 
 pytest.importorskip("tensorrt_model_connect", reason="tensorrt_model_connect requires tensorrt")
 import tensorrt_model_connect  # noqa: E402
-import tensorrt_model_connect.cli as cli  # noqa: E402
+import tensorrt_model_connect.build_cli as cli  # noqa: E402
+
+
+def test_legacy_cli_module_aliases_build_cli():
+    """Intent: keep older `tensorrt_model_connect.cli` imports working after the rename.
+    Preconditions: Both module paths are importable.
+    Postconditions: The legacy module resolves to the build_cli module object.
+    """
+    import tensorrt_model_connect.cli as compat_cli
+
+    assert compat_cli is cli
 
 
 def test_get_version_prefers_importlib_metadata():
