+++ tools/test_impact_fallback_allowlist.txt
@@ -14,7 +14,6 @@ catch_all Dockerfile.gb300 # tracked path has no precise impact rule yet; catch-
 catch_all conanfile.py # release wheel native packaging recipe; catch-all stays conservative
 catch_all conftest.py # tracked path has no precise impact rule yet; catch-all stays conservative
 catch_all docker-compose.yml # tracked path has no precise impact rule yet; catch-all stays conservative
-catch_all examples/trtmc_cli.cpp # tracked path has no precise impact rule yet; catch-all stays conservative
 catch_all examples/trtmc_dataset_benchmark.cpp # tracked path has no precise impact rule yet; catch-all stays conservative
 catch_all hf_links_wave1.txt # tracked path has no precise impact rule yet; catch-all stays conservative
 catch_all models/hf/Qwen__Qwen3-0.6B/.gitattributes # tracked path has no precise impact rule yet; catch-all stays conservative
@@ -184,6 +183,7 @@ shared_builder_module tensorrt_model_connect/pyproject.toml # shared Python buil
 shared_builder_module tensorrt_model_connect/tensorrt_model_connect/__init__.py # shared Python builder surface; broad builder impact is intentional
 shared_builder_module tensorrt_model_connect/tensorrt_model_connect/__main__.py # shared Python builder surface; broad builder impact is intentional
 shared_builder_module tensorrt_model_connect/tensorrt_model_connect/build_timing.py # shared Python builder surface; broad builder impact is intentional
+shared_builder_module tensorrt_model_connect/tensorrt_model_connect/build_cli.py # shared Python builder CLI surface; broad builder impact is intentional
 shared_builder_module tensorrt_model_connect/tensorrt_model_connect/bundle_writer.py # shared Python builder surface; broad builder impact is intentional
 shared_builder_module tensorrt_model_connect/tensorrt_model_connect/checkpoint_mapper.py # shared Python builder surface; broad builder impact is intentional
 shared_builder_module tensorrt_model_connect/tensorrt_model_connect/cli.py # shared Python builder surface; broad builder impact is intentional
