+++ website/docs/unit-design/python-builder.md
@@ -6,7 +6,7 @@ The Python builder turns a Python-first checkpoint into a native runtime bundle.
 
 ```mermaid
 flowchart TD
-  Args["trtmc build args"] --> CLI["cli.py"]
+  Args["trtmc build args"] --> CLI["build_cli.py"]
   CLI --> ConfigLayer["runtime_config CLI merge"]
   CLI --> EngineBuilder["engine_builder.py"]
   EngineBuilder --> ModelConfig["ModelConfig"]
@@ -24,7 +24,7 @@ flowchart TD
 
 ## CLI
 
-`tensorrt_model_connect/tensorrt_model_connect/cli.py` owns command parsing for `trtmc build`. It handles early `--rtx` backend selection, auto method selection, Python profile re-exec, config resolution, quantization flags, and inspection.
+`tensorrt_model_connect/tensorrt_model_connect/build_cli.py` owns command parsing for `trtmc build`. It handles early `--rtx` backend selection, auto method selection, Python profile re-exec, config resolution, quantization flags, and inspection. `tensorrt_model_connect/tensorrt_model_connect/cli.py` remains a compatibility alias for existing imports.
 
 The CLI should stay thin. It should translate user intent into builder options and leave model-specific behavior to family plugins or engine builders.
