+++ website/docs/wiki/Source-Layout.md
@@ -294,7 +294,8 @@ Recurrent model backends (Mamba, RWKV, Hybrid).
 |------|---------|
 | `__init__.py` | Package init, public `build()` API |
 | `__main__.py` | `python -m tensorrt_model_connect` entry |
-| `cli.py` | CLI: `./build/trtmc build|inspect|version` |
+| `build_cli.py` | Python builder CLI used by `trtmc build` and `python -m tensorrt_model_connect` |
+| `cli.py` | Compatibility alias for older `tensorrt_model_connect.cli` imports |
 | `config.py` | `ModelConfig` from HF `config.json` |
 | `checkpoint_mapper.py` | HF safetensors -> weight dict |
 | `graph_ops.py` | Layer 1: atomic TRT graph ops |
